+++ modypy/blocks/discrete.py
@@ -1,8 +1,9 @@
 """Blocks for discrete-time simulation"""
-from modypy.model import Block, Port, EventPort, State
+from modypy import model
+from scipy import stats
 
 
-class ZeroOrderHold(Block):
+class ZeroOrderHold(model.Block):
     """A zero-order-hold block which samples an input signal when the connected
     event occurs.
 
@@ -20,12 +21,12 @@ class ZeroOrderHold(Block):
             initial_condition: The initial state of the sampling output
                 (before the first tick of the block)
         """
-        Block.__init__(self, owner)
+        model.Block.__init__(self, owner)
 
-        self.event_input = EventPort(self)
+        self.event_input = model.EventPort(self)
         self.event_input.register_listener(self.update_state)
-        self.input = Port(shape=shape)
-        self.output = State(
+        self.input = model.Port(shape=shape)
+        self.output = model.State(
             self,
             shape=shape,
             initial_condition=initial_condition,
@@ -63,3 +64,58 @@ def zero_order_hold(system, input_port, event_port, initial_condition=None):
     hold.input.connect(input_port)
     hold.event_input.connect(event_port)
     return hold.output
+
+
+class NoiseSource(model.State):
+    """A discrete-time noise source
+
+    This noise source provides random outputs sourced from a given
+    random source.
+    The noise is sampled when the event given by the trigger occurs.
+
+    To generate band-limited white noise, use an uncorrelated random source and
+    a time constant that is sufficiently small in relation to the smallest
+    time constant in the system.
+
+    Args:
+        owner: The owner block or system
+        trigger: The event source triggering the update of the noise source.
+        shape: The shape of the output (default: ``()``)
+        random_source: A callable accepting a ``size`` argument giving the shape
+            of data to be generated
+
+    Example:
+        Collect and plot noise data:
+
+        >>> from modypy import model, simulation
+        >>> from modypy.blocks import filters, discrete
+        >>> import matplotlib.pyplot as plt
+        >>>
+        >>> system = model.System()
+        >>> noise_clock = model.Clock(owner=system, period=1/10)
+        >>> noise = discrete.NoiseSource(owner=system, trigger=noise_clock)
+        >>>
+        >>> simulator = simulation.Simulator(system=system,
+        >>>                                  start_time=0.0)
+        >>> result = simulation.SimulationResult(
+        >>>     system=system,
+        >>>     source=simulator.run_until(time_boundary=10.0))
+        >>>
+        >>> plt.plot(result.time, noise(result))
+        >>> plt.show()
+    """
+
+    def __init__(
+        self,
+        owner,
+        trigger: model.AbstractEventSource,
+        shape: model.ShapeType = (),
+        random_source=stats.norm.rvs,
+    ):
+        model.State.__init__(self, owner, shape=shape)
+        trigger.register_listener(self._update_state)
+        self.random_source = random_source
+
+    def _update_state(self, system_state):
+        """Fetch new random data"""
+        self.set_value(system_state, self.random_source(size=self.shape))